fix(config): every price in engine_config.yaml is display money.
he file said both things at once. market_maker_quotes and last_buy_price carried display money; combo legs carried raw ticks, under keys spelled the same way - and the reference comment block in every generated example documented both conventions thirty lines apart. Everything else tick-valued in the format already said so in its name (mm_max_spread_ticks, depth_snapshot_tolerance_ticks); combo legs were the only exception, in the one place where the scale is least obvious, since a combo spans instruments and two legs can sit at different tick_decimals.

+++ src/edumatcher/config_gen/builder.py
@@ -270,8 +270,9 @@ class ComboLegSpec:
     side: str
     order_type: str
     quantity: int
-    price: int | None = None
-    stop_price: int | None = None
+    #: Display money, like every price in engine_config.yaml.
+    price: float | None = None
+    stop_price: float | None = None
     smp_action: str = "NONE"
